refactor!: remove shoaling/tangs and top-level imports
