+++ pyproject.toml
@@ -24,7 +24,7 @@ multi_line_output = 5
 
 [tool.poetry]
 name = "shoal"
-version = "0.4.1"
+version = "0.5.0"
 description = "Lazy, Python-first, alternative to make, just, doit, and invoke"
 license = "MIT"
 authors = ["Kyle King <dev.act.kyle@gmail.com>"]
