+++ shoal/__init__.py
@@ -11,8 +11,3 @@ __pkg_name__ = 'shoal'
 
 # FYI: https://github.com/beartype/beartype#are-we-on-the-worst-timeline
 filterwarnings('ignore', category=BeartypeDecorHintPep585DeprecationWarning)
-
-from ._shoalling import shoalling  # noqa: E402,E408
-from ._log import configure_logger, get_logger  # noqa: E402,E408
-from ._shell import capture_shell, shell  # noqa: E402,E408
-from ._tangs import register_fun  # noqa: E402,E408
